airflow/requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-core
python-dotenv
feedparser